Refactor domain architecture, introduce RBAC, and overhaul rental workflow
- Migrated from legacy enum-based roles to a robust Claim-based Role-Based Access Control (RBAC) system using `OperationClaim` and `UserOperationClaim`.
- Redesigned the [Customer](cci:2://file:///c:/Users/MONSTER/OneDrive/Belgeler/GitHub/CarProject/Entities/Concrete/Customer.cs:6:4-19:5) entity to support both Guest and Authenticated users (`UserId` made nullable).
- Introduced `CorporateProfile` for B2B customers, linked 1-to-1 with the [Customer](cci:2://file:///c:/Users/MONSTER/OneDrive/Belgeler/GitHub/CarProject/Entities/Concrete/Customer.cs:6:4-19:5) table.
- Overhauled the [Rental](cci:2://file:///c:/Users/MONSTER/OneDrive/Belgeler/GitHub/rentacar/src/app/models/rental.ts:0:0-15:1) entity:
  - Added financial snapshot fields (`RentedDailyPrice`, `TotalPrice`, `DepositAmount`).
  - Split locations into `StartLocationId` and `EndLocationId`.
  - Incorporated `RentalStatus` (Active, Completed, Cancelled) and `DepositStatus` enums.
- Updated JWT Helper to embed user roles via `ClaimTypes.Role` for secure frontend decoding.
- Added data seeding logic and successfully applied EF Core migrations.

+++ Business/Abstract/ICarService.cs
@@ -27,7 +27,7 @@ namespace Business.Abstract
         IDataResult<CarDetailDto> GetCarDetailsById(int id);
         void CarRented(int carId);
 
-        IDataResult<int> UpdatePriceByAction(int carId, string action);
+        IDataResult<decimal> UpdatePriceByAction(int carId, string action);
 
         IDataResult<List<CarDetailDto>> GetCarsByFilters(List<int> fuelIds, List<int> gearIds, List<int> segmentIds, bool isRented, string locationName);
