+++ Core/Entities/Requests/PricingRequest.cs
@@ -3,7 +3,7 @@
     public class PricingRequest
     {
         public int CarId { get; set; }
-        public int CurrentPrice { get; set; }
+        public decimal CurrentPrice { get; set; }
         public int RentalCount { get; set; }
     }
 }
